backend/telegram-bot/requirements.txt
pyTelegramBotAPI==4.2.2
pyYAML==6.0
redis==4.0.2
